+++ src/VBox/Runtime/common/ioqueue/ioqueuebase.cpp
@@ -1,4 +1,4 @@
-/* $Id: ioqueuebase.cpp 79952 2019-07-24 11:21:03Z alexander.eichner@oracle.com $ */
+/* $Id: ioqueuebase.cpp 79983 2019-07-25 17:21:24Z alexander.eichner@oracle.com $ */
 /** @file
  * IPRT - I/O queue, Base/Public API.
  */
@@ -127,7 +127,7 @@ RTDECL(PCRTIOQUEUEPROVVTABLE) RTIoQueueProviderGetById(const char *pszId)
 
 
 RTDECL(int) RTIoQueueCreate(PRTIOQUEUE phIoQueue, PCRTIOQUEUEPROVVTABLE pProvVTable,
-                            uint32_t fFlags, size_t cSqEntries, size_t cCqEntries)
+                            uint32_t fFlags, uint32_t cSqEntries, uint32_t cCqEntries)
 {
     AssertPtrReturn(phIoQueue, VERR_INVALID_POINTER);
     AssertPtrReturn(pProvVTable, VERR_INVALID_POINTER);
@@ -141,8 +141,8 @@ RTDECL(int) RTIoQueueCreate(PRTIOQUEUE phIoQueue, PCRTIOQUEUEPROVVTABLE pProvVTa
     {
         pThis->pVTbl          = pProvVTable;
         pThis->hIoQueueProv   = (RTIOQUEUEPROV)&pThis->abInst[0];
-        pThis->cSqEntries     = (uint32_t)cSqEntries;
-        pThis->cCqEntries     = (uint32_t)cCqEntries;
+        pThis->cSqEntries     = cSqEntries;
+        pThis->cCqEntries     = cCqEntries;
         pThis->cReqsCommitted = 0;
         pThis->cReqsPrepared  = 0;
